[MOV] web: separate json controller
X-original-commit: 5ec8fce9e064e342f5df771ebe38268b3bfffae8
