+++ addons/web/controllers/__init__.py
@@ -6,6 +6,7 @@ from . import database
 from . import dataset
 from . import domain
 from . import export
+from . import json
 from . import home
 from . import model
 from . import pivot
